Fix showDisconnected handling in settings page
The old code would read pref_showDisconnected. This value wouldn't be the current value when the page is first loaded, as it's only set when the page is submitted. We therefore have to read the pref directly from the prefs. Alas upon page submission that pref would still be the old value.

Introduce the beforeRender method: "A sub called before the page is rendered, but after the prefs have been processed/saved". Exactly what we need here to set that pref.